# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UFV API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/applicators/{applicator}": {
            "get": {
                "tags": [
                    "applicators"
                ],
                "summary": "Display the specified resource.",
                "description": "show applicator",
                "operationId": "App\\Http\\Controllers\\ApplicatorController::show",
                "parameters": [
                    {
                        "name": "applicator",
                        "in": "path",
                        "description": "Applicator id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show applicator"
                    }
                }
            },
            "put": {
                "tags": [
                    "applicators"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a applicator on database",
                "operationId": "App\\Http\\Controllers\\ApplicatorController::update",
                "parameters": [
                    {
                        "name": "applicator",
                        "in": "path",
                        "description": "Applicator id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "acronym": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Applicator updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "applicators"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a applicator on database",
                "operationId": "App\\Http\\Controllers\\ApplicatorController::destroy",
                "parameters": [
                    {
                        "name": "applicator",
                        "in": "path",
                        "description": "Applicator id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Applicator deleted"
                    }
                }
            }
        },
...
        "/categories/{category}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Display the specified resource.",
                "description": "show category",
                "operationId": "App\\Http\\Controllers\\CategoryController::show",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show category"
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a category on database",
                "operationId": "App\\Http\\Controllers\\CategoryController::update",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a category on database",
                "operationId": "App\\Http\\Controllers\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    }
                }
            }
        },
...
        "/contests/{contest}": {
            "get": {
                "tags": [
                    "contests"
                ],
                "summary": "Display the specified resource.",
                "description": "show contest",
                "operationId": "App\\Http\\Controllers\\ContestController::show",
                "parameters": [
                    {
                        "name": "contest",
                        "in": "path",
                        "description": "Contest id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show contest"
                    }
                }
            },
            "put": {
                "tags": [
                    "contests"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a contest on database",
                "operationId": "App\\Http\\Controllers\\ContestController::update",
                "parameters": [
                    {
                        "name": "contest",
                        "in": "path",
                        "description": "Contest id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "institution_id": {
                                        "type": "integer"
                                    },
                                    "applicator_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contest updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "contests"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a contest on database",
                "operationId": "App\\Http\\Controllers\\ContestController::destroy",
                "parameters": [
                    {
                        "name": "contest",
                        "in": "path",
                        "description": "Contest id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contest deleted"
                    }
                }
            }
        },
...
        "/institutions/{institution}": {
            "get": {
                "tags": [
                    "institutions"
                ],
                "summary": "Display the specified resource.",
                "description": "show institution",
                "operationId": "App\\Http\\Controllers\\InstitutionController::show",
                "parameters": [
                    {
                        "name": "institution",
                        "in": "path",
                        "description": "Institution id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show institution"
                    }
                }
            },
            "put": {
                "tags": [
                    "institutions"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a institution on database",
                "operationId": "App\\Http\\Controllers\\InstitutionController::update",
                "parameters": [
                    {
                        "name": "institution",
                        "in": "path",
                        "description": "Institution id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "acronym": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Institution updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "institutions"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a institution on database",
                "operationId": "App\\Http\\Controllers\\InstitutionController::destroy",
                "parameters": [
                    {
                        "name": "institution",
                        "in": "path",
                        "description": "Institution id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Institution deleted"
                    }
                }
            }
        },
        "/questions": {
            "get": {
                "tags": [
                    "questions"
                ],
                "summary": "Display a listing of the resource",
                "description": "get all question on database and paginate then",
                "operationId": "App\\Http\\Controllers\\QuestionController::index",
                "parameters": [
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search results using key:value",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "operators",
                        "in": "query",
                        "description": "set fileds operators using field1:operator1",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "order results using field:direction",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "define page",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit per page",
                        "style": "form",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of questions"
                    }
                }
            },
            "post": {
                "tags": [
                    "questions"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "store a new question on database",
                "operationId": "App\\Http\\Controllers\\QuestionController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "statement": {
                                        "type": "string"
                                    },
                                    "subject_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "contest_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New question created"
                    }
                }
            }
        },
...
    }
}